scan for hivebins
